requirements.txt
pyaml
numpy==1.18.1
torch
opencv-python==4.5.3.56
gym[atari]==0.15.7
tensorboard
tqdm
matplotlib
dnn